style: move PtraceInterface to ptrace submodule
